remove old chart syntax attempts. rename to chart-forward-v3 to chart.ts
