+++ src/App.tsx
@@ -78,21 +78,8 @@ import { testFishRibbonChartTextured } from "./tests/fishRibbonChartTextured";
 import { testFishPolarRibbonChartTextured } from "./tests/fishPolarRibbonChartTextured";
 import { GoFishSolid } from "./ast";
 import { DitheringConfig, generateDithering } from "./tests/density";
-import { For, frame, Rect, StackX, StackY, stackY } from "./lib";
+import { For, frame, orderBy, Rect, StackX, StackY, stackY } from "./lib";
 import { testRidgeline } from "./tests/ridgeline";
-import {
-  chartBar,
-  chartFacetedBar,
-  chartFacetedBarHorizontal,
-  chartGroupedBar,
-  chartRectBF,
-  chartWaffle,
-  chartStackedBar,
-  chartArea,
-  chartSankeyIcicle,
-  chartStackedArea,
-  chartRidgeline,
-} from "./tests/chartAPITest";
 import { testOlympicMedalsStackedBars } from "./tests/olympicMedalsStackedBars";
 import { testRoseChart } from "./tests/testRoseChart";
 import { testConnectedScatterplot } from "./tests/connectedScatterplot";
@@ -100,23 +87,6 @@ import { testStringLine } from "./tests/stringline";
 import { testBumpChart } from "./tests/bump";
 import { testBoxWhiskerPlot } from "./tests/boxwhisker";
 import { testViolinPlot } from "./tests/violin";
-import {
-  v2ChartArea,
-  v2ChartBar,
-  v2ChartPolarRibbon,
-  v2ChartRect,
-  v2ChartRectSpread,
-  v2ChartRibbon,
-  v2ChartRibbonMessAround,
-  v2ChartScatter,
-  v2ChartStackedBar,
-  v2ChartPies,
-  v2ChartPie,
-  v2ChartScatterPie,
-  v2ChartMosaic,
-  v2ChartBarHorizontal,
-  // v2ChartWaffle,
-} from "./tests/chartAPITestV2";
 import {
   tailwindColorGrid,
   tailwindColorGridCompact,
@@ -127,14 +97,13 @@ import {
   tailwindColorComparison,
   tailwindColorComparisonCompact,
 } from "./tests/tailwindColors";
-import { chartForwardBar } from "./ast/marks/chart-forward";
 
 // Forward Syntax V3 imports for testing
 import {
   chart,
   spread,
-  rectForward as rect,
-  stackForward as stack,
+  rect,
+  stack,
   derive,
   layer,
   select,
@@ -148,9 +117,10 @@ import {
   normalize,
   repeat,
   scatter,
-} from "./ast/marks/chart-forward-v3";
-import { seafood } from "./data/catch";
+} from "./ast/marks/chart";
+import { catchLocations, catchLocationsArray, seafood } from "./data/catch";
 import { clock } from "./ast/coordinateTransforms/clock";
+import _ from "lodash";
 
 const ditheringTestWidth = 800;
 
@@ -359,9 +329,62 @@ const App: Component = () => {
       //     axes: true,
       //   });
       // Add your forward syntax code here:
-      chart(seafood)
-        .flow(spread("lake", { dir: "x" }), stack("species", { dir: "y" }))
-        .mark(rect({ h: "count", fill: "species" }))
+      // function bar(data, { x, y, fill }) {
+      //   return chart(data)
+      //     .flow(spread(x, { dir: "x" }))
+      //     .mark(rect({ h: y, fill }));
+      // }
+
+      /* layer({ coord: clock() }, [
+        chart(seafood)
+          .flow(
+            spread("lake", {
+              dir: "x",
+              spacing: (2 * Math.PI) / 6,
+              mode: "center",
+              y: 50,
+            }),
+            derive((d) => orderBy(d, "count", "asc")),
+            // derive((d) => normalize(d, "count")),
+            stack("species", { dir: "y", label: false })
+          )
+          .mark(rect({ h: "count", fill: "species" }))
+          .as("bars"),
+        chart(select("bars"))
+          .flow(foreach("species"))
+          .mark(area({ opacity: 0.8 })),
+      ]) */
+      /* chart(catchLocationsArray)
+        .flow(scatter("lake", { x: "x", y: "y" }))
+        .mark(circle({ r: 5 })) */
+      /* chart(seafood, { coord: clock() })
+        .flow(stack("species", { dir: "x" }))
+        .mark(rect({ w: "count", fill: "species" })) */
+      const scatterData = _(seafood)
+        .groupBy("lake")
+        .map((lakeData, lake) => ({
+          lake,
+          x: catchLocations[lake as keyof typeof catchLocations].x,
+          y: catchLocations[lake as keyof typeof catchLocations].y,
+          collection: lakeData.map((item) => ({
+            species: item.species,
+            count: item.count,
+          })),
+        }))
+        .value();
+
+      // console.log(scatterData);
+
+      // chart(seafood, { coord: clock() })
+      //   .flow(stack("species", { dir: "x" }))
+      //   .mark(rect({ w: "count", fill: "species" }))
+      chart(scatterData)
+        .flow(scatter("lake", { x: "x", y: "y" }))
+        .mark((data) =>
+          chart(data[0].collection, { coord: clock() })
+            .flow(stack("species", { dir: "x", h: 20 }))
+            .mark(rect({ w: "count", fill: "species" }))
+        )
         .render(testContainer, {
           w: 400,
           h: 400,
@@ -407,19 +430,10 @@ const App: Component = () => {
       <GoFishSolid w={500} h={300} defs={defs} axes={true}>
         {testBoxWhiskerPlot()}
       </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartBar()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartBarHorizontal()}
-      </GoFishSolid>
       <GoFishSolid w={350} h={300} defs={defs}>
         {testMosaic()}
       </GoFishSolid>
       <br />
-      <GoFishSolid w={320} h={400} defs={defs} axes={false}>
-        {v2ChartMosaic()}
-      </GoFishSolid>
       <h1>Tailwind Color Palette Variants</h1>
 
       <h2>Color Comparison (Compact)</h2>
@@ -467,39 +481,6 @@ const App: Component = () => {
       </div>
       <br />
       <br />
-      <GoFishSolid w={500} h={300} x={200} y={200} defs={defs} axes={false}>
-        {v2ChartPie()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={false}>
-        {v2ChartScatterPie()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartRibbonMessAround()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={false}>
-        {v2ChartScatter()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={false}>
-        {v2ChartRect()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={false}>
-        {v2ChartRectSpread()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartBar()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartStackedBar()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartRibbon()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} x={150} y={150} defs={defs} axes={false}>
-        {v2ChartPolarRibbon()}
-      </GoFishSolid>
-      <GoFishSolid w={320} h={400} defs={defs} axes={true}>
-        {v2ChartArea()}
-      </GoFishSolid>
       <GoFishSolid w={500} h={500} defs={defs}>
         {StackX(
           { spacing: 60 },
@@ -530,15 +511,6 @@ const App: Component = () => {
             )
           )
         )}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={500} defs={defs}>
-        {chartArea()}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={500} defs={defs}>
-        {chartStackedArea()}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={500} defs={defs}>
-        {chartRidgeline()}
       </GoFishSolid>
       <GoFishSolid w={500} h={300} defs={defs}>
         {testViolinPlot()}
@@ -571,9 +543,6 @@ const App: Component = () => {
         {testFishStackedBar()}
       </GoFishSolid>
       <GoFishSolid w={500} h={500} defs={defs}>
-        {chartSankeyIcicle()}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={500} defs={defs}>
         {testSankeyIcicleAPIv2()}
       </GoFishSolid>
       <GoFishSolid w={500} h={500} defs={defs}>
@@ -586,27 +555,6 @@ const App: Component = () => {
         {testFishWaffleAPIv2()}
       </GoFishSolid>
       <GoFishSolid w={500} h={500} defs={defs}>
-        {chartWaffle()}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={250} defs={defs}>
-        {chartFacetedBarHorizontal()}
-      </GoFishSolid>
-      <GoFishSolid w={250} h={500} defs={defs}>
-        {chartFacetedBar()}
-      </GoFishSolid>
-      <GoFishSolid w={500} h={300} defs={defs}>
-        {chartGroupedBar()}
-      </GoFishSolid>
-      <GoFishSolid w={250} h={300} defs={defs}>
-        {chartStackedBar()}
-      </GoFishSolid>
-      <GoFishSolid w={250} h={300} defs={defs}>
-        {chartBar()}
-      </GoFishSolid>
-      <GoFishSolid w={250} h={300} defs={defs}>
-        {chartRectBF()}
-      </GoFishSolid>
-      <GoFishSolid w={250} h={300} defs={defs}>
         {testRidgeline()}
       </GoFishSolid>
       <GoFishSolid w={250} h={300} defs={defs}>
